// src/types/TemplatesProps.ts
// le typage

export interface TemplateProps {
  Lastname: string;
  firstName: string;
  jobTitle: string;
  Departement: string;
  company: string;
  // combiner le duex nums
  Officephone: string;
  MobilPhone: string;
  website: string;
  email: string;
  Adresse: string;
  // socialLinks
  Instagram: string;
  facebook: string;
  twitter: string;
  LinkedIn: string;
  WhatsApp: string;
  // images
  PersonnalImage: string;
  LogoImage: string;
  HandWrittingImage: string;
}